Update #50 Navigation Visualizer Additions,
Now the Debug Navigation fuction shows the built path after its done, and the A* has been re-weighted to be a bit more directional, but may be reverted to help with pathfinding

+++ src/com/GlitchyDev/World/Navigation/NavigatingEntity.java
@@ -1,5 +1,5 @@
 package com.GlitchyDev.World.Navigation;
 
 public interface NavigatingEntity {
-    int getMovementCost(PathType pathType);
+    double getMovementCost(PathType pathType);
 }
